requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
sqlalchemy
psycopg2-binary==2.9.6
websockets
